pyproject.toml
[tool.poetry]
name = "xmasdraw"
version = "0.1.0"
description = ""
authors = ["Nautik <nautik1@n-p.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
PyYAML = "^6.0.1"
pyuwsgi = "^2.0.23"
argon2-cffi = "^23.1.0"
marshmallow = "^3.20.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
